Fix Coverity issues in shared memory, VT-d ACPI, and L1TF masks
Issue Type: Unintended sign extension
- Cast sender_id to uint32_t before left shift in FF-A SMC return path.

Issue Type: Out-of-bounds access
- Validate DMAR table length before parsing.
- Use validated table length for loop bounds.
- Zero the DMAR table from its base pointer with bounded length.

Issue Type: Overflowed constant
- Replace (-1U) shift-mask patterns in L1TF mask derivation with an
explicit full-width uint32_t mask.
- Avoid UINT32_MAX so the low-level EVMM build does not depend on an
undeclared standard macro.

Tracked-On: NIACP3-1383

+++ vmm/modules/spm_srv/shared-mem-smcall.c
@@ -361,7 +361,7 @@ long trusty_ffa_fill_desc(guest_cpu_handle_t gcpu,
 	if (obj->desc_filled != obj->desc_size) {
 		SMC_RET8(gcpu, smc_handle, SMC_FC_FFA_MEM_FRAG_RX, handle_low,
 				handle_high, obj->desc_filled,
-				obj->desc.sender_id << 16, 0, 0, 0);
+				((uint32_t)obj->desc.sender_id) << 16U, 0, 0, 0);
 	}
 
 	SMC_RET8(gcpu, smc_handle, SMC_FC_FFA_SUCCESS, 0, handle_low, handle_high, 0,
